pyproject.toml
[project]
name = "progetto-calcio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "beautifulsoup4>=4.12.3",
    "geopandas>=1.0.1",
    "matplotlib-colors>=1.0.16",
    "matplotlib>=3.10.3",
    "numpy>=2.2.0",
    "polars>=1.17.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "streamlit>=1.41.0",
]
